+++ packages/terra-action-footer/CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 ### Changed
 * Styling improvements
 * Component markup refactoring
+* Restructured theme tests into a separate file
 
 1.27.0 - (October 24, 2018)
 ------------------